+++ src/lib/log.js
@@ -1,4 +1,4 @@
-import config from './config';
+import config from '../../config/config';
 
 const logLevels = [
 	'debug',
@@ -7,18 +7,26 @@ const logLevels = [
 	'error'
 ];
 
-export default logLevels.reduce((logger, funcName, index) => {
-	logger[funcName] = function (...args) {
-		const consoleFunc = funcName === 'debug' ? 'log' : funcName;
-		const { logging } = config;
-		if (logging && console && typeof console[consoleFunc] === 'function') {
-			const enabledLevelIndex = logLevels.indexOf(logging.toString().toLocaleLowerCase());
-			if (logging === true || (enabledLevelIndex > -1 && index >= enabledLevelIndex)) {
-				const [message, ...rest] = [...args];
-				console[consoleFunc](`${funcName.toUpperCase()} - (CMP) ${message}`, ...rest);
-			}
-		}
-	};
-	return logger;
-}, {});
+class Log {
+	constructor(logLevel) {
+		this.logLevel = logLevel || config.logLevel || 'warn';
+
+		logLevels.reduce((logger, funcName, index) => {
+			logger[funcName] = function (...args) {
+				const consoleFunc = funcName === 'debug' ? 'log' : funcName;
+				if (this.logLevel && console && typeof console[consoleFunc] === 'function') {
+					const enabledLevelIndex = logLevels.indexOf(this.logLevel.toString().toLocaleLowerCase());
+					if (enabledLevelIndex > -1 && index >= enabledLevelIndex) {
+						const [message, ...rest] = [...args];
+						console[consoleFunc](`${funcName.toUpperCase()} - (CMP) ${message}`, ...rest);
+					}
+				}
+			};
+			return logger;
+		}, this);
+	}
+}
+
+const log = new Log();
+export default log;