+++ webpack.config.babel.js
@@ -103,6 +103,7 @@ module.exports = [
 		entry: {
 			'cmp.ndmtag': ['core-js/fn/promise', './cmp.ndmtag.js'],
 			'cmp.stub': './cmp.stub.js',
+			'cmp.dfp': './cmp.dfp.js',
 		},
 
 		output: {