+++ CHANGELOG.md
@@ -9,8 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Replace our own CMP with the QuantCast CMP
 
 ### Removed
-- Removed support for the custom CMP implementation
-- Removed `ndmtag.bundle.js` (not to be confused with `cmp.ndmtag.bundle.js`).
+- Remove support for the custom CMP implementation (`cmp.custom.bundle.js`)
+- Remove `ndmtag.bundle.js` (not to be confused with `cmp.ndmtag.bundle.js`)
+- Remove `cmp.ssp.bundle.js`
 
 ## [1.3.1] - 2019-08-06
 ### Changed