+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add script `cmp.dfp.bundle.js` to render AppNexus tags with DFP.
+
 ### Changed
 - Replace our own CMP with the QuantCast CMP
 - Read the configuration from `config/config.js`.