+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [2.0.0-beta.1] - 2019-09-06
 ### Added
 - Add script `cmp.dfp.bundle.js` to render AppNexus tags with DFP.
 
@@ -101,3 +103,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [1.3.0-beta.3]: https://github.com/nextdaymedia/cmp/tree/1.3.0-beta.3
 [1.3.0]: https://github.com/nextdaymedia/cmp/tree/1.3.0
 [1.3.1]: https://github.com/nextdaymedia/cmp/tree/1.3.1
+[2.0.0-beta.1]: https://github.com/nextdaymedia/cmp/tree/2.0.0-beta.1